Moved contexts to own folder & disabled User Context
